day04/requirements.txt
biopython>=1.81
certifi>=2023.7.22
